Added reset functionality, cleaned up CSS.
Added functionality to the reset button, and had the page display
default values upon loading.

I also gave the reset button a max width and the same CSS classes
as the other apps, so that when I include this page in the rest of
my landing page it will be consistent.